pyproject.toml
[project]
name = "multi-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langgraph-cli[inmem]>=0.3.6",
    "langgraph-supervisor>=0.0.29",
    "loguru>=0.7.3",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "langchain[openai]>=0.3.27",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src", "src.*"]
